トップレベルの処理をindex.jsに切り出した
